+++ requests_oauthlib/compliance_fixes/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+
+from .facebook import facebook_compliance_fix