+++ HISTORY.rst
@@ -4,3 +4,4 @@ History
 v0.4.0dev
 +++++++++
 - OAuth1Session methods only return unicode strings. #55.
+- Added facebook compliance fix and access_token_response hook to OAuth2Session. #63.